BaseHelmUpdater no longer inherits from BaseUpdater
